requirements.txt
simplejson
requests
ndg-httpsclient
gitpython
